refactor(Force_Field): extract subsystems into systems/ subdirectory (#715)
Moves the four subsystem modules into src/games/Force_Field/src/systems/
without changing any game logic:

- systems/combat_system.py   — weapon firing, hit detection, damage
- systems/spawn_manager.py   — Force Field-specific spawn manager
- systems/entity_manager.py  — bot and projectile management
- systems/audio_manager.py   — new; owns sound wiring via the event bus

Top-level combat_system.py, entity_manager.py, and spawn_manager.py
become thin backward-compatibility re-exports so existing tests continue
to pass without modification.

game.py now imports directly from systems/ and constructs AudioManager
to own the event-bus audio subscription (previously inline in
_wire_event_bus).

Closes #715
