+++ src/games/Force_Field/src/systems/__init__.py
@@ -1,0 +1,15 @@
+"""Force Field subsystem modules."""
+
+from __future__ import annotations
+
+from .audio_manager import AudioManager as AudioManager
+from .combat_system import CombatSystem as CombatSystem
+from .entity_manager import EntityManager as EntityManager
+from .spawn_manager import FFSpawnManager as FFSpawnManager
+
+__all__ = [
+    "AudioManager",
+    "CombatSystem",
+    "EntityManager",
+    "FFSpawnManager",
+]
