+++ src/games/Force_Field/src/systems/combat_system.py
@@ -1,0 +1,611 @@
+from __future__ import annotations
+
+import logging
+import math
+import random
+from typing import TYPE_CHECKING
+
+import pygame
+
+from games.shared.constants import COMBO_TIMER_FRAMES
+
+from .. import constants as C  # noqa: N812
+from ..projectile import Projectile
+
+if TYPE_CHECKING:
+    from ..bot import Bot
+    from ..game import Game
+    from ..player import Player
+
+logger = logging.getLogger(__name__)
+
+
+class CombatSystem:
+    """Handles weapon firing, hit detection, and damage application."""
+
+    def __init__(self, game: Game):
+        self.game = game
+
+    @property
+    def player(self) -> Player:
+        """Return the active player, raising ValueError if none exists."""
+        if self.game.player is None:
+            raise ValueError("Player is None")
+        return self.game.player
+
+    _FIRE_DISPATCH: dict[str, str] = {
+        "hitscan": "_fire_hitscan",
+        "projectile": "_fire_projectile",
+        "spread": "_fire_spread",
+        "beam": "_fire_beam",
+        "burst": "_fire_burst",
+    }
+
+    _RECOIL: dict[str, float] = {
+        "minigun": 0.5,
+        "shotgun": 4.0,
+        "rocket": 5.0,
+    }
+
+    def fire_weapon(self, is_secondary: bool = False) -> None:
+        """Handle weapon firing via data-driven dispatch."""
+        weapon_name = self.player.current_weapon
+        weapon_data = C.WEAPONS[weapon_name]
+
+        # Recoil Effect
+        if self.player.zoomed:
+            recoil_amount = 0.5
+        else:
+            recoil_amount = self._RECOIL.get(weapon_name, 2.0)
+        self.player.pitch += recoil_amount
+
+        # Sound
+        self.game.sound_manager.play_sound(f"shoot_{weapon_name}")
+
+        if is_secondary:
+            self.check_shot_hit(is_secondary=True)
+            return
+
+        fire_mode = weapon_data.get("fire_mode", "hitscan")
+        handler = getattr(self, self._FIRE_DISPATCH[fire_mode])
+        handler(weapon_data, weapon_name)
+
+    def _fire_hitscan(self, weapon_data: dict, weapon_name: str) -> None:
+        """Fire a single hitscan ray."""
+        self.check_shot_hit(is_secondary=False)
+
+    def _fire_projectile(self, weapon_data: dict, weapon_name: str) -> None:
+        """Spawn a projectile, delegating to weapon-specific handlers."""
+        _projectile_handlers = {
+            "plasma": self._fire_plasma,
+            "pulse": self._fire_pulse,
+            "rocket": self._fire_rocket,
+            "flamethrower": self._fire_flamethrower,
+            "freezer": self._fire_freezer,
+        }
+        handler = _projectile_handlers.get(weapon_name)
+        if handler:
+            handler()
+        else:
+            self._fire_generic_projectile(weapon_data, weapon_name)
+
+    def _fire_generic_projectile(self, weapon_data: dict, weapon_name: str) -> None:
+        """Fallback projectile spawner for weapons without a specific handler."""
+        p = Projectile(
+            self.player.x,
+            self.player.y,
+            self.player.angle,
+            speed=float(weapon_data.get("projectile_speed", 0.5)),
+            damage=self.player.get_current_weapon_damage(),
+            is_player=True,
+            color=weapon_data.get("projectile_color", (0, 255, 255)),
+            size=0.3,
+            weapon_type=weapon_name,
+        )
+        self.game.entity_manager.add_projectile(p)
+
+    def _fire_spread(self, weapon_data: dict, weapon_name: str) -> None:
+        """Fire multiple pellets with spread."""
+        pellets = int(weapon_data.get("pellets", 8))
+        spread = float(weapon_data.get("spread", 0.15))
+        for _ in range(pellets):
+            angle_off = random.uniform(-spread, spread)
+            self.check_shot_hit(angle_offset=angle_off)
+
+    def _fire_beam(self, weapon_data: dict, weapon_name: str) -> None:
+        """Fire a hitscan beam with visual trail."""
+        self.check_shot_hit(is_secondary=False, is_laser=True)
+
+    def _fire_burst(self, weapon_data: dict, weapon_name: str) -> None:
+        """Fire a burst of fast projectiles via minigun handler."""
+        self._fire_minigun()
+
+    def _fire_plasma(self) -> None:
+        p = Projectile(
+            self.player.x,
+            self.player.y,
+            self.player.angle,
+            speed=float(C.WEAPONS["plasma"].get("projectile_speed", 0.5)),
+            damage=self.player.get_current_weapon_damage(),
+            is_player=True,
+            color=C.WEAPONS["plasma"].get("projectile_color", (0, 255, 255)),
+            size=0.225,
+            weapon_type="plasma",
+        )
+        self.game.entity_manager.add_projectile(p)
+
+    def _fire_pulse(self) -> None:
+        p = Projectile(
+            self.player.x,
+            self.player.y,
+            self.player.angle,
+            speed=float(C.WEAPONS["pulse"].get("projectile_speed", 0.6)),
+            damage=self.player.get_current_weapon_damage(),
+            is_player=True,
+            color=C.WEAPONS["pulse"].get("projectile_color", (100, 100, 255)),
+            size=0.3,
+            weapon_type="pulse",
+        )
+        self.game.entity_manager.add_projectile(p)
+
+    def _fire_rocket(self) -> None:
+        p = Projectile(
+            self.player.x,
+            self.player.y,
+            self.player.angle,
+            speed=float(C.WEAPONS["rocket"].get("projectile_speed", 0.3)),
+            damage=self.player.get_current_weapon_damage(),
+            is_player=True,
+            color=C.WEAPONS["rocket"].get("projectile_color", (255, 100, 0)),
+            size=0.6,
+            weapon_type="rocket",
+        )
+        self.game.entity_manager.add_projectile(p)
+
+    def _fire_flamethrower(self) -> None:
+        damage = self.player.get_current_weapon_damage()
+        # Spawn multiple flame particles
+        for _ in range(2):
+            angle_off = random.uniform(-0.15, 0.15)
+            final_angle = self.player.angle + angle_off
+            speed = float(C.WEAPONS["flamethrower"].get("projectile_speed", 0.35))
+
+            # Add slight speed variation
+            speed *= random.uniform(0.8, 1.2)
+
+            p = Projectile(
+                self.player.x,
+                self.player.y,
+                final_angle,
+                damage,
+                speed=speed,
+                is_player=True,
+                color=C.WEAPONS["flamethrower"].get("projectile_color", (255, 140, 0)),
+                size=0.4,
+                weapon_type="flamethrower",
+            )
+            self.game.entity_manager.add_projectile(p)
+
+    def _fire_freezer(self) -> None:
+        damage = self.player.get_current_weapon_damage()
+        angle_off = random.uniform(-0.05, 0.05)
+        final_angle = self.player.angle + angle_off
+        speed = float(C.WEAPONS["freezer"].get("projectile_speed", 0.4))
+
+        p = Projectile(
+            self.player.x,
+            self.player.y,
+            final_angle,
+            damage,
+            speed=speed,
+            is_player=True,
+            color=C.WEAPONS["freezer"].get("projectile_color", (200, 255, 255)),
+            size=0.25,
+            weapon_type="freezer",
+        )
+        self.game.entity_manager.add_projectile(p)
+
+    def _fire_minigun(self) -> None:
+        damage = self.player.get_current_weapon_damage()
+        num_bullets = 3
+        for _ in range(num_bullets):
+            angle_off = random.uniform(-0.15, 0.15)
+            final_angle = self.player.angle + angle_off
+
+            p = Projectile(
+                self.player.x,
+                self.player.y,
+                final_angle,
+                damage,
+                speed=2.0,
+                is_player=True,
+                color=(255, 255, 0),
+                size=0.1,
+                weapon_type="minigun",
+            )
+            self.game.entity_manager.add_projectile(p)
+
+        self.game.particle_system.add_explosion(
+            C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2, count=8, color=(255, 255, 0)
+        )
+
+    def check_shot_hit(
+        self,
+        is_secondary: bool = False,
+        angle_offset: float = 0.0,
+        is_laser: bool = False,
+    ) -> None:
+        """Check if player's shot hit a bot"""
+        if not (self.game.raycaster is not None):
+            raise ValueError("DbC Blocked: Precondition failed.")
+
+        weapon_range = self.player.get_current_weapon_range()
+        if is_secondary:
+            weapon_range = 100
+
+        weapon_damage = self.player.get_current_weapon_damage()
+
+        current_spread = C.SPREAD_ZOOM if self.player.zoomed else C.SPREAD_BASE
+        if angle_offset == 0.0:
+            spread_offset = random.uniform(-current_spread, current_spread)
+            aim_angle = self.player.angle + spread_offset
+        else:
+            aim_angle = self.player.angle + angle_offset
+
+        aim_angle %= 2 * math.pi
+
+        wall_dist, _, _, _, _ = self.game.raycaster.cast_ray(
+            self.player.x, self.player.y, aim_angle
+        )
+
+        if wall_dist > weapon_range:
+            wall_dist = float(weapon_range)
+
+        closest_bot = None
+        closest_dist = float("inf")
+        is_headshot = False
+
+        wall_dist_sq = wall_dist * wall_dist
+
+        for bot in self.game.bots:
+            if not bot.alive:
+                continue
+
+            dx = bot.x - self.player.x
+            dy = bot.y - self.player.y
+            dist_sq = dx * dx + dy * dy
+
+            if dist_sq > wall_dist_sq:
+                continue
+
+            distance = math.sqrt(dist_sq)
+
+            bot_angle = math.atan2(dy, dx)
+            angle = bot_angle if bot_angle >= 0 else bot_angle + 2 * math.pi
+            bot_angle_norm = angle
+
+            angle_diff = abs(bot_angle_norm - aim_angle)
+            if angle_diff > math.pi:
+                angle_diff = 2 * math.pi - angle_diff
+
+            if angle_diff < 0.15:
+                if distance < closest_dist:
+                    closest_bot = bot
+                    closest_dist = distance
+                    is_headshot = angle_diff < C.HEADSHOT_THRESHOLD
+
+        if is_secondary:
+            self._handle_secondary_hit(closest_bot, closest_dist, wall_dist)
+            return
+
+        if is_laser:
+            self.game.particle_system.add_laser(
+                start=(C.SCREEN_WIDTH - 200, C.SCREEN_HEIGHT - 180),
+                end=(C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2),
+                color=(255, 0, 0),
+                timer=5,
+                width=3,
+            )
+
+        if closest_bot:
+            self._apply_damage(
+                closest_bot, closest_dist, weapon_range, weapon_damage, is_headshot
+            )
+
+    def _handle_secondary_hit(
+        self, closest_bot: Bot | None, closest_dist: float, wall_dist: float
+    ) -> None:
+        impact_dist = wall_dist
+        if closest_bot and closest_dist < wall_dist:
+            impact_dist = closest_dist
+
+        ray_angle = self.player.angle
+        impact_x = self.player.x + math.cos(ray_angle) * impact_dist
+        impact_y = self.player.y + math.sin(ray_angle) * impact_dist
+
+        try:
+            self.explode_laser(impact_x, impact_y)
+        except (ValueError, AttributeError, pygame.error):
+            logger.exception("Error in explode_laser")
+
+        self.game.particle_system.add_laser(
+            start=(C.SCREEN_WIDTH - 200, C.SCREEN_HEIGHT - 180),
+            end=(C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2),
+            color=(0, 255, 255),
+            timer=C.LASER_DURATION,
+            width=C.LASER_WIDTH,
+        )
+
+    def _apply_damage(
+        self,
+        bot: Bot,
+        distance: float,
+        weapon_range: float,
+        base_damage: int,
+        is_headshot: bool,
+    ) -> None:
+        range_factor = max(0.3, 1.0 - (distance / weapon_range))
+
+        # Re-calculate angle diff for accuracy factor
+        dx = bot.x - self.player.x
+        dy = bot.y - self.player.y
+        bot_angle = math.atan2(dy, dx)
+        angle = bot_angle if bot_angle >= 0 else bot_angle + 2 * math.pi
+
+        angle_diff = abs(angle - self.player.angle)
+        if angle_diff > math.pi:
+            angle_diff = 2 * math.pi - angle_diff
+
+        accuracy_factor = max(0.5, 1.0 - (angle_diff / 0.15))
+
+        final_damage = int(base_damage * range_factor * accuracy_factor)
+
+        bot.take_damage(final_damage, is_headshot=is_headshot)
+
+        if self.game.show_damage:
+            self.game.damage_texts.append(
+                {
+                    "x": C.SCREEN_WIDTH // 2 + random.randint(-20, 20),
+                    "y": C.SCREEN_HEIGHT // 2 - 50,
+                    "text": str(final_damage) + ("!" if is_headshot else ""),
+                    "color": C.RED if is_headshot else C.DAMAGE_TEXT_COLOR,
+                    "timer": 60,
+                    "vy": -1.0,
+                }
+            )
+
+        self.game.particle_system.add_explosion(
+            C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2, count=5
+        )
+
+        # Improved Blood Feedback
+        blood_color = (200, 0, 0)
+        # For non-standard enemies (like robots/aliens), maybe different blood?
+        if bot.enemy_type in ["ball", "minigunner"]:  # Maybe robots?
+            blood_color = (20, 20, 20)  # Oil?
+        elif bot.enemy_type == "demon":
+            blood_color = (0, 200, 0)  # Green slime
+
+        for _ in range(15):
+            speed = random.uniform(2, 8)
+            angle = random.uniform(0, 2 * math.pi)
+            self.game.particle_system.add_particle(
+                x=C.SCREEN_WIDTH // 2,
+                y=C.SCREEN_HEIGHT // 2,
+                dx=math.cos(angle) * speed,
+                dy=math.sin(angle) * speed,
+                color=blood_color,
+                timer=random.randint(20, 40),
+                size=random.randint(3, 6),
+                gravity=0.2,  # Gravity for blood
+            )
+
+        if not bot.alive:
+            self._handle_kill(bot)
+
+    def _handle_kill(self, bot: Bot) -> None:
+        self.game.kills += 1
+        self.game.kill_combo_count += 1
+        self.game.kill_combo_timer = COMBO_TIMER_FRAMES
+        self.game.last_death_pos = (bot.x, bot.y)
+        self.game.sound_manager.play_sound("scream")
+        self.game.event_bus.emit("bot_killed", x=bot.x, y=bot.y)
+
+    def explode_bomb(self, projectile: Projectile) -> None:
+        """Handle bomb explosion logic"""
+        dist_to_player = math.sqrt(
+            (projectile.x - self.player.x) ** 2 + (projectile.y - self.player.y) ** 2
+        )
+
+        try:
+            self.game.sound_manager.play_sound("bomb")
+        except (pygame.error, OSError):
+            logger.exception("Bomb Audio Failed")
+
+        # Screen shake on bomb
+        if dist_to_player < 30:
+            shake = max(0, 30 - dist_to_player)
+            self.game.screen_shake = shake
+
+        if dist_to_player < 20:
+            if dist_to_player < 10:
+                self.game.particle_system.add_particle(
+                    x=C.SCREEN_WIDTH // 2,
+                    y=C.SCREEN_HEIGHT // 2,
+                    dx=0,
+                    dy=0,
+                    color=C.WHITE,
+                    timer=20,
+                    size=3000,
+                )
+
+            explosion_center = (C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2)
+
+            for _ in range(100):
+                angle = random.uniform(0, 2 * math.pi)
+                speed = random.uniform(5, 20)
+                color = random.choice([C.ORANGE, C.RED, C.YELLOW, (50, 50, 50)])
+                self.game.particle_system.add_particle(
+                    x=explosion_center[0],
+                    y=explosion_center[1],
+                    dx=math.cos(angle) * speed,
+                    dy=math.sin(angle) * speed,
+                    color=color,
+                    timer=random.randint(40, 80),
+                    size=random.randint(5, 15),
+                )
+
+        for bot in self.game.bots:
+            if not bot.alive:
+                continue
+            dx = bot.x - projectile.x
+            dy = bot.y - projectile.y
+            dist = math.sqrt(dx * dx + dy * dy)
+
+            if dist < C.BOMB_RADIUS:
+                damage_factor = 1.0 - (dist / C.BOMB_RADIUS)
+                damage = int(1000 * damage_factor)
+
+                if bot.take_damage(damage):
+                    self._handle_kill(bot)
+
+                if dist < 5.0 and dist_to_player < 20:
+                    self.game.particle_system.add_explosion(
+                        C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2, count=3
+                    )
+
+        self.game.damage_texts.append(
+            {
+                "x": C.SCREEN_WIDTH // 2,
+                "y": C.SCREEN_HEIGHT // 2 - 100,
+                "text": "BOOM!",
+                "color": C.ORANGE,
+                "timer": 90,
+                "vy": -0.5,
+            }
+        )
+
+    def explode_laser(self, impact_x: float, impact_y: float) -> None:
+        """Trigger Massive Laser Explosion at Impact Point"""
+        try:
+            self.game.sound_manager.play_sound("boom_real")
+        except (pygame.error, OSError):
+            logger.exception("Boom sound failed")
+
+        self.game.screen_shake = 10.0
+
+        hits = 0
+        for bot in self.game.bots:
+            if not hasattr(bot, "alive"):
+                continue
+
+            if bot.alive:
+                dist = math.sqrt((bot.x - impact_x) ** 2 + (bot.y - impact_y) ** 2)
+                if dist < C.LASER_AOE_RADIUS:
+                    damage = 500
+                    killed = bot.take_damage(damage)
+                    hits += 1
+
+                    if killed:
+                        self._handle_kill(bot)
+
+                    for _ in range(10):
+                        self.game.particle_system.add_particle(
+                            x=C.SCREEN_WIDTH // 2,
+                            y=C.SCREEN_HEIGHT // 2,
+                            dx=random.uniform(-10, 10),
+                            dy=random.uniform(-10, 10),
+                            color=(
+                                random.randint(200, 255),
+                                0,
+                                random.randint(200, 255),
+                            ),
+                            timer=40,
+                            size=random.randint(4, 8),
+                        )
+
+        if hits > 0:
+            self.game.damage_texts.append(
+                {
+                    "x": C.SCREEN_WIDTH // 2,
+                    "y": C.SCREEN_HEIGHT // 2 - 80,
+                    "text": "LASER ANNIHILATION!",
+                    "color": (255, 0, 255),
+                    "timer": 60,
+                    "vy": -2,
+                }
+            )
+
+    def _explode_generic(
+        self, projectile: Projectile, radius: float, weapon_type: str
+    ) -> None:
+        """Generic explosion logic"""
+        dist_to_player = math.sqrt(
+            (projectile.x - self.player.x) ** 2 + (projectile.y - self.player.y) ** 2
+        )
+        if dist_to_player < 15:
+            self.game.damage_flash_timer = 15
+            self.game.screen_shake = 15.0 - dist_to_player
+
+        if dist_to_player < 20:
+            count = 5 if weapon_type == "rocket" else 3
+            self.game.particle_system.add_explosion(
+                C.SCREEN_WIDTH // 2, C.SCREEN_HEIGHT // 2, count=count
+            )
+
+            color = C.CYAN if weapon_type == "plasma" else C.ORANGE
+            if weapon_type == "pulse":
+                color = (100, 100, 255)
+            elif weapon_type == "freezer":
+                color = (200, 255, 255)
+
+            for _ in range(20):
+                self.game.particle_system.add_particle(
+                    x=C.SCREEN_WIDTH // 2,
+                    y=C.SCREEN_HEIGHT // 2,
+                    dx=random.uniform(-10, 10),
+                    dy=random.uniform(-10, 10),
+                    color=color,
+                    timer=40,
+                    size=random.randint(4, 8),
+                )
+
+        try:
+            self.game.sound_manager.play_sound(
+                "boom_real" if weapon_type == "rocket" else "shoot_plasma"
+            )
+        except (pygame.error, OSError):
+            pass
+
+        for bot in self.game.bots:
+            if not bot.alive:
+                continue
+            dx = bot.x - projectile.x
+            dy = bot.y - projectile.y
+            dist = math.sqrt(dx * dx + dy * dy)
+
+            if dist < radius:
+                damage_factor = 1.0 - (dist / radius)
+                damage = int(projectile.damage * damage_factor)
+
+                if bot.take_damage(damage):
+                    self._handle_kill(bot)
+                elif weapon_type == "freezer":
+                    bot.freeze(180)  # Freeze for 3 seconds
+
+    def explode_plasma(self, projectile: Projectile) -> None:
+        """Trigger a plasma AoE explosion at the projectile's position."""
+        self._explode_generic(projectile, C.PLASMA_AOE_RADIUS, "plasma")
+
+    def explode_pulse(self, projectile: Projectile) -> None:
+        """Trigger a pulse AoE explosion at the projectile's position."""
+        self._explode_generic(projectile, C.PULSE_AOE_RADIUS, "pulse")
+
+    def explode_rocket(self, projectile: Projectile) -> None:
+        """Trigger a rocket AoE explosion using the configured blast radius."""
+        radius = float(C.WEAPONS["rocket"].get("aoe_radius", 6.0))
+        self._explode_generic(projectile, radius, "rocket")
+
+    def explode_freezer(self, projectile: Projectile) -> None:
+        """Trigger a freezer AoE explosion, slowing enemies within range."""
+        self._explode_generic(projectile, 4.0, "freezer")  # Custom radius for freezer
